Create Graph Object
Transformed variable into object so that the script can be used independent of the original page.  However, this does make the original page quite a bit slower since each graph is it's own object and therefore the path for each graph is calculated separately.  This leads me to think that completing work on making the script function independently is falling outside the scope of the original task and should be developed separately.