Completely reworked logic for path that draws graph
Fixed a few bugs and introduced a few.  Graphs no longer extend beyond
the graphing window.  Graphs incorrectly approximate the xVal
corresponding the the min or max yVal when a graph is leaving or
entering the defined window.  Extremely fast growth may not be graphed
if a value does not appear within the window (530(x+.5)).  Shading does
not use the updated graphing logic and will have to be reworked
separately.